+++ backend/src/services/service.services.ts
@@ -1,6 +1,7 @@
 import Service from "@entities/Service";
 import ServiceRepository from "@repositories/service.repository";
 import { DeleteResult, UpdateResult } from "typeorm";
+import { validateNumberId } from "@functionality/validation";
 
 export const createServiceService = async (data: Partial<Service>): Promise<Service> => {
     return await ServiceRepository.createService(data);
@@ -11,6 +12,8 @@ export const readAllServicesService = async (): Promise<Service[]> => {
 }
 
 export const readServiceByIdService = async (id: number): Promise<Service> => {
+    if (!validateNumberId(id)) throw new Error("invalid-id");
+
     return await ServiceRepository.readServiceById(id)
 }
 
@@ -19,5 +22,7 @@ export const updateServiceService = async (data: Partial<Service>): Promise<Upda
 }
 
 export const deleteServiceService = async (id: number): Promise<DeleteResult> => {
+    if (!validateNumberId(id)) throw new Error("invalid-id");
+
     return await ServiceRepository.deleteService(id);
 }
